+++ kibot/misc.py
@@ -452,6 +452,7 @@ FONT_HELP_TEXT = ('\n        Important: If you use custom fonts and/or colors pl
                   'global variable.')
 MULTI_SCH_NAME = ('\nNote that only for KiCad 10 and newer the `output` option controls the names of all sub-sheets.'
                   '\nIn this case %f and %F contains the project and the sub-sheet path inside the hierarchy')
+SUBIMAGE_REGEX = r'_([^\[]+)(?:\[(\d+)\])?'
 # CSS style for HTML tables used by BoM and E/DRC
 # Light colors
 BG_GEN = "#DCF5E4"